server/requirements.txt
flask
redis
flask-cors
flask_pymongo
flask_restful
flask-expects-json
pyjwt